Add timeout to queue item acquisition
This improves parallel processing and allows the software not to block on any occasion. Right now, if some queues are still empty and yet not intended for use and a thread tries to acquire for an item in that queue, it stays blocking till the queue gets at least one item. Due to fair queue acquisition, this can happen in another round and one after another, threads can get to a deadlock waiting for an item.